Settings: add QuantumDroid version and Build date
Every ROM claims it as their own and in turn making it very difficult to
figure out who the real author is.

It might of originated with CM so credit to them if so but I'm not entirely sure
who exactly within the CM org.

If someone knows, hit me up and we'll give them a gold star and a hand shake!

Signed-off-by: Hriday Sharma <hridaysharma42@gmail.com>